+++ CHANGES.md
@@ -19,6 +19,7 @@
 - FIX: Fixing DeprecationWarning `invalid escape sequence \.`
 - FIX: Computing Brightness Temperature of `Landsat` TIR bands instead of leaving them as is
 - FIX: Better management of Landsat Instrument values
+- FIX: Better radiometry attribute (adding `brightness temperature` and `reflectance and brightness temperature` values)
 - FIX: Changing `Brilliance Temperature` to the correct `Brightness Temperature`
 
 ### Optimizations